refactor(cloudinary delete): modified code for image delete in cloudinary
